+++ lib/src/data_source/mod.rs
@@ -1,11 +1,10 @@
 #[cfg(feature = "plugin")]
 pub mod connector_plugin;
-pub mod csv_file;
+pub mod csv;
 #[cfg(feature = "flight")]
 pub mod flight_stream;
 pub mod infer_schema;
-pub mod json_file;
-pub mod json_rest;
+pub mod json;
 pub mod location_uri;
 pub mod nd_json_file;
 pub mod nd_json_rest;