+++ lib/src/request/mod.rs
@@ -1,3 +1,4 @@
 pub mod body;
+pub mod custom_extractor;
 pub mod header;
 pub mod query_param;