+++ lib/src/response/handler/data_source.rs
@@ -2,16 +2,16 @@
 // Sasaki, Naoki <nsasaki@sal.co.jp> January 15, 2023
 //
 
+use std::str::FromStr;
+use std::sync::Arc;
+
+use axum::{extract, http::StatusCode, response::IntoResponse};
+use serde::Serialize;
+
 use crate::context::session_manager::SessionManager;
 use crate::data_source::schema::DataSourceSchema;
 use crate::request::body::DataSources;
 use crate::response::http_error::ResponseError;
-use axum::extract;
-use axum::extract::Path;
-use axum::http::StatusCode;
-use axum::response::IntoResponse;
-use serde::Serialize;
-use std::sync::Arc;
 
 #[derive(Serialize)]
 #[allow(clippy::module_name_repetitions)]
@@ -23,7 +23,7 @@ pub struct DataSourceDetail {
 
 pub async fn index<S: SessionManager>(
     extract::State(session_mgr): extract::State<Arc<tokio::sync::Mutex<S>>>,
-    Path(session_id): Path<String>,
+    extract::Path(session_id): extract::Path<String>,
 ) -> Result<impl IntoResponse, ResponseError> {
     log::info!("Accessing index of session data sources responder");
 
@@ -36,7 +36,7 @@ pub async fn index<S: SessionManager>(
 
 pub async fn detail<S: SessionManager>(
     extract::State(session_mgr): extract::State<Arc<tokio::sync::Mutex<S>>>,
-    Path((session_id, name)): Path<(String, String)>,
+    extract::Path((session_id, name)): extract::Path<(String, String)>,
 ) -> Result<impl IntoResponse, ResponseError> {
     log::info!("Accessing session data source detail responder");
 
@@ -51,7 +51,7 @@ pub async fn detail<S: SessionManager>(
 
 pub async fn create<S: SessionManager>(
     extract::State(session_mgr): extract::State<Arc<tokio::sync::Mutex<S>>>,
-    Path(session_id): Path<String>,
+    extract::Path(session_id): extract::Path<String>,
     extract::Json(payload): extract::Json<DataSources>,
 ) -> Result<impl IntoResponse, ResponseError> {
     log::info!("Accessing register data sources to session context handler");
@@ -65,9 +65,71 @@ pub async fn create<S: SessionManager>(
     Ok(StatusCode::NO_CONTENT)
 }
 
+pub async fn upload<S: SessionManager>(
+    extract::State(session_mgr): extract::State<Arc<tokio::sync::Mutex<S>>>,
+    extract::Path(session_id): extract::Path<String>,
+    mut multipart: extract::Multipart,
+) -> Result<impl IntoResponse, ResponseError> {
+    while let Some(mut field) = multipart
+        .next_field()
+        .await
+        .map_err(|e| ResponseError::request_validation(format!("Invalid multipart content: {e}")))?
+    {
+        let content_type = mime::Mime::from_str(field.content_type().ok_or(
+            ResponseError::request_validation("The content-type must be specified"),
+        )?)
+        .unwrap();
+
+        let name = field
+            .name()
+            .map(String::from)
+            .ok_or(ResponseError::request_validation(
+                "The name of field is must required",
+            ))?;
+
+        let mut bytes_buffer = bytes::BytesMut::new();
+
+        while let Some(chunk) = field.chunk().await.map_err(|e| {
+            ResponseError::payload_too_large(format!(
+                "Invalid multipart chunk, May be exceeds payload size limit: {e}"
+            ))
+        })? {
+            bytes_buffer.extend_from_slice(&chunk);
+        }
+
+        let locked_session_mgr = session_mgr.lock().await;
+
+        match (content_type.type_(), content_type.subtype().as_str()) {
+            (mime::APPLICATION, "vnd.apache.parquet") => {
+                locked_session_mgr
+                    .append_parquet_bytes(&session_id, &name, bytes_buffer.freeze())
+                    .await?;
+            }
+            (mime::APPLICATION, "json") => {
+                locked_session_mgr
+                    .append_json_bytes(&session_id, &name, bytes_buffer.freeze())
+                    .await?;
+            }
+            (mime::TEXT, "csv") => {
+                locked_session_mgr
+                    .append_csv_bytes(&session_id, &name, bytes_buffer.freeze())
+                    .await?;
+            }
+            _ => {
+                return Err(ResponseError::unsupported_format(
+                    "The content-type of the multipart field must be either \
+                            'text/csv', 'application/json', or 'application/vnd.apache.parquet'",
+                ));
+            }
+        }
+    }
+
+    Ok(StatusCode::NO_CONTENT)
+}
+
 pub async fn save<S: SessionManager>(
     extract::State(session_mgr): extract::State<Arc<tokio::sync::Mutex<S>>>,
-    Path(session_id): Path<String>,
+    extract::Path(session_id): extract::Path<String>,
     extract::Json(payload): extract::Json<DataSources>,
 ) -> Result<impl IntoResponse, ResponseError> {
     log::info!("Accessing save data sources to local file handler");
@@ -83,9 +145,9 @@ pub async fn save<S: SessionManager>(
 
 pub async fn remove<S: SessionManager>(
     extract::State(session_mgr): extract::State<Arc<tokio::sync::Mutex<S>>>,
-    Path((session_id, name)): Path<(String, String)>,
+    extract::Path((session_id, name)): extract::Path<(String, String)>,
 ) -> Result<impl IntoResponse, ResponseError> {
-    log::info!("Accessing data source remove handler");
+    log::info!("Accessing remove data source handler");
 
     session_mgr
         .lock()
@@ -98,9 +160,9 @@ pub async fn remove<S: SessionManager>(
 
 pub async fn refresh<S: SessionManager>(
     extract::State(session_mgr): extract::State<Arc<tokio::sync::Mutex<S>>>,
-    Path((session_id, name)): Path<(String, String)>,
+    extract::Path((session_id, name)): extract::Path<(String, String)>,
 ) -> Result<impl IntoResponse, ResponseError> {
-    log::info!("Accessing data source refresh handler");
+    log::info!("Accessing refresh data sources handler");
 
     session_mgr
         .lock()